Moose twinning data for Tetlin National Wildlife Refuge

A dataset containing observations of moose during moose twinning surveys at National Wildlife Refuge (https://iris.fws.gov/APPS/ServCat/Reference/Profile/179584). The variables are as follows:

Usage

moose_twinning

Format

A data frame with 501 rows and 14 variables:

srvy_day

date of the survey (2021-05-29–2024-05-31)

leafout

percent emergence of leaves for deciduous shrubs and trees (60–75)

group

group number or waypoint from field form

latitude

Y coordinate: Decimal degrees to 6 decimal places; WGS84

longitude

X coordinate: Decimal degrees to 6 decimal places; WGS84

bull

number of bulls observed in group (0–5)

yrlcow

number of yearling cows observed in group (0–3)

cow0clf

number of cows with no calves observed in group (0–5)

cow1clf

number of cows with a single calf observed in group (0–2)

cow2clf

number of cows with twins observed in group (0–1)

cow3clf

number of cows with three calves observed in group (0–0)

pilot

first and last name of the pilot

observer

first and last name of the observer

nabevw

name of the survey sub-unit (East, West)